Rename fastdns to edgedns w/backwards compat for now
